update: update the alover code

+++ iris_prediction/serializers.py
@@ -2,7 +2,7 @@ from rest_framework import serializers
 
 
 class IrisSerializer(serializers.Serializer):
-    sepal_length = serializers.IntegerField()
-    sepal_width = serializers.IntegerField()
-    petal_length = serializers.IntegerField()
-    petal_width = serializers.IntegerField()
+    sepal_length = serializers.FloatField()
+    sepal_width = serializers.FloatField()
+    petal_length = serializers.FloatField()
+    petal_width = serializers.FloatField()
